trace: Allow response mutation in `OnResponse`
Sometimes it is useful to be able to mutate the response to add extra
headers based on the trace context, which is currently not possible as
the response it not mutable in the trait definition of `OnResponse`.

This commit changes that so that users of the library can do things like
adding trace ids into the headers of responses.